refactor(OMN-11575, OMN-11580): rename omnimemory directories + split observability
- utils/ -> adapters/ with adapter_* prefix per ONEX naming convention
- observability.py split into adapter_correlation_context.py, adapter_metrics_counter.py, adapter_observability_wrapper.py
- clients/ -> adapters/ in node_kreuzberg_parse_effect, node_memory_retrieval_effect, node_agent_learning_retrieval_effect
- client_kreuzberg.py -> adapter_kreuzberg_client.py; embedding_client.py -> adapter_embedding_client.py
- tools/stubs/ -> tools/validators/; contract_validator.py -> validator_contract.py; ProtocolContractValidator -> ValidatorContract
- utils/ in node_intent_event_consumer_effect and node_intent_query_effect -> mappers/
- Update io_audit_whitelist.yaml, transport_import_whitelist.yaml, and validate_http_imports.py for new paths
- All imports updated; 2733 tests pass, pre-commit clean
