+++ src/omnimemory/nodes/node_intent_event_consumer_effect/mappers/__init__.py
@@ -3,7 +3,7 @@
 
 """Utilities for intent event consumer effect node."""
 
-from omnimemory.nodes.node_intent_event_consumer_effect.utils.util_event_mapper import (
+from omnimemory.nodes.node_intent_event_consumer_effect.mappers.util_event_mapper import (
     map_event_to_storage_request,
 )
