+++ src/omnimemory/nodes/node_kreuzberg_parse_effect/adapters/__init__.py
