+++ src/omnimemory/nodes/node_memory_lifecycle_orchestrator/handlers/handler_memory_expire.py
@@ -104,7 +104,7 @@ from omnimemory.enums import EnumLifecycleState
 
 if TYPE_CHECKING:
     from omnibase_core.container import ModelONEXContainer
-from omnimemory.utils.concurrency import (
+from omnimemory.adapters.adapter_concurrency import (
     CircuitBreaker,
     CircuitBreakerOpenError,
     CircuitBreakerState,
