+++ src/omnimemory/nodes/node_memory_retrieval_effect/adapters/adapter_embedding_client.py
@@ -14,7 +14,7 @@ Example::
 
     import asyncio
     import os
-    from omnimemory.nodes.node_memory_retrieval_effect.clients import (
+    from omnimemory.nodes.node_memory_retrieval_effect.adapters import (
         EmbeddingClient,
         ModelEmbeddingClientConfig,
     )
