+++ tests/audit/io_audit_whitelist.yaml
@@ -39,12 +39,12 @@ files:
       - "net-client"
       - "env-access"
 
-  - path: "src/omnimemory/nodes/node_memory_retrieval_effect/clients/embedding_client.py"
+  - path: "src/omnimemory/nodes/node_memory_retrieval_effect/adapters/adapter_embedding_client.py"
     reason: "Effect client uses httpx for embedding API calls; expected I/O boundary"
     allowed_rules:
       - "net-client"
 
-  - path: "src/omnimemory/nodes/node_kreuzberg_parse_effect/clients/client_kreuzberg.py"
+  - path: "src/omnimemory/nodes/node_kreuzberg_parse_effect/adapters/adapter_kreuzberg_client.py"
     reason: "Kreuzberg effect client: httpx for API calls, Path.read_text/write_text for idempotency cache
       (OMN-2733)"
     allowed_rules:
