+++ tests/audit/transport_import_whitelist.yaml
@@ -22,7 +22,7 @@ schema_version: "1.0.0"
 
 files:
   # Infrastructure utilities - health checks require direct connectivity probes
-  - path: "src/omnimemory/utils/health_manager.py"
+  - path: "src/omnimemory/adapters/adapter_health_manager.py"
     reason: "Health manager probes asyncpg and redis connectivity directly; Phase 3 will abstract behind
       protocol"
     allowed_modules:
@@ -47,12 +47,12 @@ files:
       - asyncpg
 
   # Effect node clients - effect nodes are the designated I/O boundary
-  - path: "src/omnimemory/nodes/node_memory_retrieval_effect/clients/embedding_client.py"
+  - path: "src/omnimemory/nodes/node_memory_retrieval_effect/adapters/adapter_embedding_client.py"
     reason: "Effect client uses httpx for embedding API calls; expected I/O boundary"
     allowed_modules:
       - httpx
 
-  - path: "src/omnimemory/nodes/node_kreuzberg_parse_effect/clients/client_kreuzberg.py"
+  - path: "src/omnimemory/nodes/node_kreuzberg_parse_effect/adapters/adapter_kreuzberg_client.py"
     reason: "Kreuzberg effect client uses httpx for document extraction API calls; expected I/O boundary
       (OMN-2733)"
     allowed_modules:
