+++ tests/integration/nodes/test_kreuzberg_parse_effect.py
@@ -33,7 +33,7 @@ import pytest
 from omnimemory.models.crawl.model_document_discovered_event import (
     ModelDocumentDiscoveredEvent,
 )
-from omnimemory.nodes.node_kreuzberg_parse_effect.clients.client_kreuzberg import (
+from omnimemory.nodes.node_kreuzberg_parse_effect.adapters.adapter_kreuzberg_client import (
     KreuzbergExtractionError,
     KreuzbergExtractResult,
 )
