+++ tests/unit/nodes/test_kreuzberg_parse_effect/test_handler_kreuzberg_parse.py
@@ -33,7 +33,7 @@ import pytest
 from omnimemory.models.crawl.model_document_discovered_event import (
     ModelDocumentDiscoveredEvent,
 )
-from omnimemory.nodes.node_kreuzberg_parse_effect.clients.client_kreuzberg import (
+from omnimemory.nodes.node_kreuzberg_parse_effect.adapters.adapter_kreuzberg_client import (
     KreuzbergExtractionError,
     KreuzbergExtractResult,
     KreuzbergTimeoutError,
@@ -54,7 +54,9 @@ _FAKE_CONTENT_BLOB_REF = "sha256:" + "a" * 64
 
 # Module paths for patching — patch where names are looked up, not where defined
 _HANDLER_MOD = "omnimemory.nodes.node_kreuzberg_parse_effect.handler_kreuzberg_parse"
-_CLIENT_MOD = "omnimemory.nodes.node_kreuzberg_parse_effect.clients.client_kreuzberg"
+_CLIENT_MOD = (
+    "omnimemory.nodes.node_kreuzberg_parse_effect.adapters.adapter_kreuzberg_client"
+)
 
 
 # ---------------------------------------------------------------------------
